+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-* Feed manager. Support to API FeedCount method 
+* [`Order`] Added `setOrderItems` method
+* [FeedManager](docs/Managers/Feed.md). Support to API `FeedCount` method 
 
 ### Changed