+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * [`Order`] Added `setOrderItems` method
+* [FeedManager](docs/Managers/Feed.md). Support to API `FeedCount` method 
 
 ### Changed