+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+### Fixed
+
+## [0.4.0] - 2020-08-31
+
+### Added
 * [`Order`] Added `setOrderItems` method
 * [[FeedManager](docs/Managers/Feed.md)] Support to API `FeedCount` method 
 * [[FeedManager](docs/Managers/Feed.md)] Support [FeedOffsetList](https://developers.linio.com/reference/feedoffsetlist) action
@@ -80,7 +88,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial Release
 
-[Unreleased]: https://github.com/LinioIT/seller-center-sdk/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/LinioIT/seller-center-sdk/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/LinioIT/seller-center-sdk/releases/tag/v0.4.0
 [0.3.0]: https://github.com/LinioIT/seller-center-sdk/releases/tag/v0.3.0
 [0.2.1]: https://github.com/LinioIT/seller-center-sdk/releases/tag/v0.2.1
 [0.2.0]: https://github.com/LinioIT/seller-center-sdk/releases/tag/v0.2.0